简易网盘后端-更新文件秒传功能

+++ CloudPan/src/main/java/com/lms/cloudpan/task/DeleteFilesTask.java
@@ -18,6 +18,7 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.List;
+import java.util.stream.Collectors;
 
 //定期删除有删除标记的文件或者文件夹
 @Component
@@ -31,21 +32,38 @@ public class DeleteFilesTask {
         IFileService fileService = SpringUtil.getBean(IFileService.class);
         OssProperties ossProperties = SpringUtil.getBean(OssProperties.class);
         IFolderService folderService = SpringUtil.getBean(IFolderService.class);
-        OssClient ossClient = SpringUtil.getBean(OssClient.class);
+
+        //获取未删除的文件信息fingerPrint字段
+        List<String> fingerPrints = fileService.list(new QueryWrapper<File>().eq("delete_flag", 0))
+                .stream().map(File::getFingerPrint).collect(Collectors.toList());
+
+
+
+
         List<File> deleteFlag = fileService.list(new QueryWrapper<File>().eq("delete_flag", 1));
         //删除实际文件
         deleteFlag.forEach(file -> {
-            //http://localhost:9998/pan/static/bucket_user_11/2023/05/22/f10a9aeb-b76f-4777-bc4c-1001e6e3b4c3.png
-            String[] split=file.getFileUrl().split(FileConstants.STATIC_REQUEST_PREFIX);
-            String realPath=ossProperties.getRootPath()+split[1];
-            Path path = Paths.get(realPath);
-            if(Files.exists(path)){
-                try {
-                    Files.delete(path);
-                } catch (IOException e) {
-                    throw new RuntimeException(e);
+
+
+            //如果其他用户还在使用就不需要删除
+            long count = fingerPrints.stream().filter(fingerPrint -> file.getFingerPrint().equals(fingerPrint))
+                    .count();
+            //只有没有其他人使用这个文件的时候才删除实际文件
+            if(count<=0){
+                //http://localhost:9998/pan/static/bucket_user_11/2023/05/22/f10a9aeb-b76f-4777-bc4c-1001e6e3b4c3.png
+                String[] split=file.getFileUrl().split(FileConstants.STATIC_REQUEST_PREFIX);
+                String realPath=ossProperties.getRootPath()+split[1];
+                Path path = Paths.get(realPath);
+                if(Files.exists(path)){
+                    try {
+                        Files.delete(path);
+                    } catch (IOException e) {
+                        throw new RuntimeException(e);
+                    }
                 }
             }
+
+
             fileService.removeById(file.getFileId());
         });
         //删除文件夹记录
